chore: Move vct package out of activitypub
Move vct package out of activitypub

Closes #1267
