+++ cmd/orb-server/startcmd/start.go
@@ -81,10 +81,6 @@ import (
 	"github.com/trustbloc/orb/pkg/activitypub/service/activityhandler"
 	"github.com/trustbloc/orb/pkg/activitypub/service/anchorsynctask"
 	apspi "github.com/trustbloc/orb/pkg/activitypub/service/spi"
-	"github.com/trustbloc/orb/pkg/activitypub/service/vct"
-	"github.com/trustbloc/orb/pkg/activitypub/service/vct/logmonitoring"
-	"github.com/trustbloc/orb/pkg/activitypub/service/vct/logmonitoring/handler"
-	"github.com/trustbloc/orb/pkg/activitypub/service/vct/proofmonitoring"
 	apariesstore "github.com/trustbloc/orb/pkg/activitypub/store/ariesstore"
 	apmemstore "github.com/trustbloc/orb/pkg/activitypub/store/memstore"
 	activitypubspi "github.com/trustbloc/orb/pkg/activitypub/store/spi"
@@ -143,6 +139,10 @@ import (
 	"github.com/trustbloc/orb/pkg/store/wrapper"
 	"github.com/trustbloc/orb/pkg/taskmgr"
 	"github.com/trustbloc/orb/pkg/vcsigner"
+	"github.com/trustbloc/orb/pkg/vct"
+	"github.com/trustbloc/orb/pkg/vct/logmonitoring"
+	"github.com/trustbloc/orb/pkg/vct/logmonitoring/handler"
+	"github.com/trustbloc/orb/pkg/vct/proofmonitoring"
 	"github.com/trustbloc/orb/pkg/webcas"
 	wfclient "github.com/trustbloc/orb/pkg/webfinger/client"
 )
