+++ pkg/vct/logmonitoring/verifier/verifier.go
